mod rvba wflow+script

+++ rvba/build-rvba-app.sh
@@ -16,20 +16,6 @@ docker rm -vf $(docker ps -aq)
 
 log "Cloning termux-app"
 git clone --depth=1 --no-tags https://github.com/termux/termux-app
-
-#  log "Patching setup scripts"
-#  sed -i '/venv/d' termux-packages/scripts/setup-ubuntu.sh
-#  sed -i 's/openjdk-18/openjdk-17/g' termux-packages/scripts/setup-ubuntu.sh
-#  sed -i 's#cmdline-tools/bin#cmdline-tools/latest/bin#g' termux-packages/scripts/setup-android-sdk.sh
-
-#  log "Running setup scripts"
-#  ./termux-packages/scripts/setup-ubuntu.sh
-#  ./termux-packages/scripts/setup-android-sdk.sh
-
-#  log "Restoring the setup scripts"
-#  git reset HEAD termux-packages/scripts/setup-ubuntu.sh
-#  git reset HEAD termux-packages/scripts/setup-android-sdk.sh
-
 cd termux-app
 
 log "Changing package IDs"
@@ -39,11 +25,13 @@ git grep -l 'com_termux' | xargs sed -i 's#com_termux#com_reisxd_rvba#g'
 log "Patching downloadBootstrap()"
 sed -i '#def downloadBootstrap#a return;' app/build.gradle
 
-log "Copying aarch64 bootstrap"
-cp ~/bootstrap-aarch64.zip app/src/main/cpp
+if [[ $1 == aarch64 ]] || [[ -z $1 ]]; then
 
-log "Patching termux-bootstrap-zip.S"
-cat <<EOF >app/src/main/cpp/termux-bootstrap-zip.S
+  log "Copying aarch64 bootstrap"
+  cp ~/bootstrap-aarch64.zip app/src/main/cpp
+
+  log "Patching termux-bootstrap-zip.S"
+  cat <<EOF >app/src/main/cpp/termux-bootstrap-zip.S
 asm
    .global blob
    .global blob_size
@@ -65,21 +53,23 @@ asm
    .int 1b - blob
 EOF
 
-log "Patching app/build.gradle"
-sed "s#include 'x86', 'x86_64', 'armeabi-v7a', 'arm64-v8a'#include 'arm64-v8a'#g" app/build.gradle
+  log "Patching app/build.gradle"
+  sed "s#include 'x86', 'x86_64', 'armeabi-v7a', 'arm64-v8a'#include 'arm64-v8a'#g" app/build.gradle
 
-log "Patching terminal-emulator/build.gradle"
-sed "s#abiFilters 'x86', 'x86_64', 'armeabi-v7a', 'arm64-v8a'#abiFilters 'arm64-v8a'#g"
+  log "Patching terminal-emulator/build.gradle"
+  sed "s#abiFilters 'x86', 'x86_64', 'armeabi-v7a', 'arm64-v8a'#abiFilters 'arm64-v8a'#g"
 
-log "Building app for aarch64"
-./gradlew -a --daemon --parallel --build-cache --configuration-cache build
+  log "Building app for aarch64"
+  ./gradlew -a --daemon --parallel --build-cache --configuration-cache build
 
-log "Copying arm bootstrap"
-rm app/src/main/cpp/bootstrap-aarch64.zip
-cp ~/bootstrap-arm.zip app/src/main/cpp
+elif [[ $1 == arm ]]; then
 
-log "Patching termux-bootstrap-zip.S"
-cat <<EOF >app/src/main/cpp/termux-bootstrap-zip.S
+  log "Copying arm bootstrap"
+  rm app/src/main/cpp/bootstrap-aarch64.zip
+  cp ~/bootstrap-arm.zip app/src/main/cpp
+
+  log "Patching termux-bootstrap-zip.S"
+  cat <<EOF >app/src/main/cpp/termux-bootstrap-zip.S
 asm
    .global blob
    .global blob_size
@@ -102,11 +92,16 @@ asm
    .int 1b - blob
 EOF
 
-log "Patching app/build.gradle"
-sed "s#include 'arm64-v8a'#include 'armeabi-v7a'#g" app/build.gradle
+  log "Patching app/build.gradle"
+  sed "s#include 'arm64-v8a'#include 'armeabi-v7a'#g" app/build.gradle
 
-log "Patching terminal-emulator/build.gradle"
-sed "s#abiFilters 'arm64-v8a'#abiFilters 'armeabi-v7a'#g" terminal-emulator/build.gradle
+  log "Patching terminal-emulator/build.gradle"
+  sed "s#abiFilters 'arm64-v8a'#abiFilters 'armeabi-v7a'#g" terminal-emulator/build.gradle
 
-log "Building app for arm"
-./gradlew -a --daemon --parallel --build-cache --configuration-cache build
+  log "Building app for arm"
+  ./gradlew -a --daemon --parallel --build-cache --configuration-cache build
+
+else
+  log "Invalid arch: $1"
+  exit 128
+fi
